fix: make gradle plus conan work
