+++ ai_playground/src/main/AndroidManifest.xml
