+++ conan/profiles/android_common
@@ -1,0 +1,18 @@
+[settings]
+os=Android
+os.api_level={{os.getenv("ANDROID_API_V")}}
+compiler=clang
+compiler.cppstd=gnu17
+compiler.version=17
+compiler.libcxx=c++_static
+build_type=Release
+
+[conf]
+tools.android:ndk_path={{os.getenv("ANDROID_NDK_HOME")}}
+tools.cmake.cmaketoolchain:extra_variables={"CMAKE_WARN_DEPRECATED":"OFF","CMAKE_ERROR_DEPRECATED":"OFF","CMAKE_SUPPRESS_DEVELOPER_WARNINGS":"ON"}
+
+[buildenv]
+PATH+=(path){{os.getenv("ANDROID_NDK_HOME")}}/toolchains/llvm/prebuilt/linux-x86_64/bin
+
+[tool_requires] 
+cmake/3.31.5
